cli/create/build tests ok
