+++ programs/cli/spec/fixtures/chatgpt/README.md
@@ -1,4 +1,4 @@
-# my-chatgpt
+# chatgpt
 
 > This project was bootstrapped using the Extension.js React-TypeScript template.
