+++ programs/cli/spec/fixtures/chatgpt/package.json
